make data folder and move datas
