+++ src/components/Header/header.module.css
@@ -17,7 +17,7 @@
 }
 .btn {
     position: fixed;
-    left: 1660px;
+    right: 100px;
     top: 10px;
     font-size: 16px;
     background-color: rgba(117, 139, 255, 0);
@@ -26,7 +26,7 @@
 
 .line {
     position: fixed;
-    left: 1656px;
+    right: 200px;
     top: 5px;
     font-size: 22px;
     font-weight: 900;
